requirements.txt
openai>=1.0.0
pyyaml>=6.0.1
aiohttp>=3.8.0
python-dotenv>=1.0.0
networkx>=3.2.1
graphviz>=0.20.1
